refactor(web): counterfactual
