+++ apps/web/src/components/dashboard/FirstSteps/index.tsx
@@ -4,7 +4,7 @@ import ExternalLink from '@/components/common/ExternalLink'
 import ModalDialog from '@/components/common/ModalDialog'
 import QRCode from '@/components/common/QRCode'
 import Track from '@/components/common/Track'
-import { FirstTxFlow, ActivateAccountButton } from '@/features/counterfactual/components'
+import { FirstTxFlow, ActivateAccountButton } from '@/features/counterfactual'
 import { selectUndeployedSafe } from '@/features/counterfactual/store'
 import { isReplayedSafeProps } from '@/features/counterfactual/services'
 import useBalances from '@/hooks/useBalances'
