+++ apps/web/src/components/sidebar/NewTxButton/index.tsx
@@ -1,5 +1,4 @@
-import { ActivateAccountButton } from '@/features/counterfactual/components'
-import { useIsCounterfactualSafe } from '@/features/counterfactual'
+import { ActivateAccountButton, useIsCounterfactualSafe } from '@/features/counterfactual'
 import { type ReactElement, useContext } from 'react'
 import Button from '@mui/material/Button'
 import { OVERVIEW_EVENTS, trackEvent, MixpanelEventParams } from '@/services/analytics'
