+++ apps/web/src/components/sidebar/SidebarHeader/index.tsx
@@ -1,4 +1,4 @@
-import { CounterfactualStatusButton } from '@/features/counterfactual/components'
+import { CounterfactualStatusButton } from '@/features/counterfactual'
 import { type ReactElement } from 'react'
 import IconButton from '@mui/material/IconButton'
 import Tooltip from '@mui/material/Tooltip'
