+++ apps/web/src/components/tx-flow/actions/Counterfactual.tsx
@@ -1,8 +1,7 @@
 import { SafeTxContext } from '@/components/tx-flow/SafeTxProvider'
 import { useCallback, useContext } from 'react'
 import { TxFlowContext } from '../TxFlowProvider'
-import { CounterfactualForm } from '@/features/counterfactual/components'
-import { useIsCounterfactualSafe } from '@/features/counterfactual'
+import { CounterfactualForm, useIsCounterfactualSafe } from '@/features/counterfactual'
 import { type SlotComponentProps, SlotName, withSlot } from '../slots'
 
 const Counterfactual = ({ onSubmitSuccess }: SlotComponentProps<SlotName.Submit>) => {
