+++ apps/web/src/features/counterfactual/components/CounterfactualStatusButton/index.tsx
@@ -2,30 +2,10 @@ import { selectUndeployedSafe } from '../../store/undeployedSafesSlice'
 import useSafeInfo from '@/hooks/useSafeInfo'
 import InfoIcon from '@/public/images/notifications/info.svg'
 import { useAppSelector } from '@/store'
-import AutorenewRoundedIcon from '@mui/icons-material/AutorenewRounded'
-import { IconButton, Tooltip, type SvgIconProps } from '@mui/material'
+import SpinningIcon from '@/components/common/SpinningIcon'
+import { IconButton, Tooltip } from '@mui/material'
 import classnames from 'classnames'
 import css from './styles.module.css'
-
-export const LoopIcon = (props: SvgIconProps) => {
-  return (
-    <AutorenewRoundedIcon
-      {...props}
-      sx={{
-        ...props.sx,
-        animation: 'spin 2s linear infinite',
-        '@keyframes spin': {
-          '0%': {
-            transform: 'rotate(0)',
-          },
-          '100%': {
-            transform: 'rotate(360deg)',
-          },
-        },
-      }}
-    />
-  )
-}
 
 const CounterfactualStatusButton = () => {
   const { safe, safeAddress } = useSafeInfo()
@@ -48,7 +28,7 @@ const CounterfactualStatusButton = () => {
         color={isActivating ? 'info' : 'warning'}
         disableRipple
       >
-        {isActivating ? <LoopIcon /> : <InfoIcon />}
+        {isActivating ? <SpinningIcon /> : <InfoIcon />}
       </IconButton>
     </Tooltip>
   )
