--- apps/web/src/features/counterfactual/components/index.ts
@@ -1,10 +1,0 @@
-export { default as ActivateAccountButton } from './ActivateAccountButton'
-export { default as ActivateAccountFlow } from './ActivateAccountFlow'
-export { default as CheckBalance } from './CheckBalance'
-export { default as CounterfactualForm } from './CounterfactualForm'
-export { default as CounterfactualHooks } from './CounterfactualHooks'
-export { default as CounterfactualStatusButton, LoopIcon } from './CounterfactualStatusButton'
-export { default as CounterfactualSuccessScreen } from './CounterfactualSuccessScreen'
-export { default as FirstTxFlow } from './FirstTxFlow'
-export { default as LazyCounterfactual } from './LazyCounterfactual'
-export { default as PayNowPayLater } from './PayNowPayLater'
